api/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
pydantic
tzdata
python-dotenv
ortools
httpx
pytest

